university-project/back/src/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.19.0
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
boto3==1.34.2
botocore==1.34.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
distro==1.9.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.2
fastapi-storages==0.2.1
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
fastapi_camelcase==2.0.0
frozenlist==1.4.0
greenlet==3.0.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
makefun==1.15.1
MarkupSafe==2.1.3
multidict==6.0.4
openai==1.6.1
outcome==1.3.0.post0
packaging==23.2
paramiko==3.2.0
passlib==1.7.4
Pillow==10.1.0
psycopg2-binary==2.9.9
pyaes==1.6.1
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyhumps==3.8.0
PyJWT==2.8.0
PyNaCl==1.5.0
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
s3transfer==0.9.0
schedule==1.2.1
selenium==4.14.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqladmin==0.15.2
SQLAlchemy==2.0.22
SQLAlchemy-Utc==0.14.0
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
text-unidecode==1.3
tiktoken==0.5.2
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
Unidecode==1.3.7
urllib3==1.26.16
uvicorn==0.23.2
watchfiles==0.21.0
webdriver-manager==4.0.1
websockets==11.0.3
wsproto==1.2.0
WTForms==3.0.1
yarl==1.9.2
